# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nemesis API",
    "version": "1.0.0",
    "description": "Artists platform API - Manage artworks, events, and artist profiles"
  },
  "servers": [
    {
      "url": "http://localhost:5005",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/cart/add": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "artworkId": {
                    "example": "any"
                  },
                  "eventId": {
                    "example": "any"
                  },
                  "quantity": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/update": {
      "patch": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "artworkId": {
                    "example": "any"
                  },
                  "eventId": {
                    "example": "any"
                  },
                  "quantity": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/remove/{itemId}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/geocode/": {
      "post": {
        "tags": [
          "Geocode"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "street": {
                    "example": "any"
                  },
                  "streetNum": {
                    "example": "any"
                  },
                  "zipCode": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/geocode/reverse": {
      "post": {
        "tags": [
          "Geocode"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lat": {
                    "example": "any"
                  },
                  "lng": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authToken",
        "description": "JWT token stored in HTTP-only cookie"
      }
    },
...
  }
}